MidiDev/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# MidiDev/midi2vol.py: 6
pulsectl == 20.5.1

# MidiDev/midi2vol.py: 7
pyalsaaudio == 0.9.0

# MidiDev/midi2vol.py: 3
python_rtmidi == 1.4.3
